modules/location-gulp/requirements.txt
kafka-python==2.0.2
grpcio-tools==1.51.3
Faker==17.6.0